Extracted EventFormState.ts
